Meta-graph: create virtual partitions
First, the graph matrix is symmetric, so BuildFragments produces only
one half of the edges: (P.X):(Q.Y) is created iff P < Q (obviously
P is NEVER equal to Q).
Then, create a "virtual node" V for each edge so that (P.X):(Q.Y) is
represented by to edges:
  * From one node representing P to V
  * From V to one node representing Q

However, the edges there should be "free" (i.e. distance equal 0), but
this is another topic, since we also need to weight connections to
external partitions in the sub-graphs.